refactor(core): remove fs service abstraction
