+++ packages/core/src/fs/index.ts
@@ -1,13 +1,9 @@
 export { FileWatchService, type FileWatchServiceOptions } from './file-watch-service';
-export { FsService } from './fs-service';
 export { realpathOrResolve } from './paths';
 export type {
-  FileReadResult,
-  FileStat,
   FileChangeKind,
   FileWatchOptions,
   IFileWatchService,
-  IFsService,
   RawFileEvent,
   WatchHandle,
 } from './types';
