+++ packages/core/src/fs/types.ts
@@ -33,31 +33,3 @@ export type IFileWatchService = {
   ): WatchHandle;
   dispose(): Promise<void>;
 };
-
-export type FileReadResult = {
-  content: string;
-  truncated: boolean;
-  totalSize: number;
-};
-
-export type FileStat = {
-  type: 'file' | 'dir';
-  size: number;
-  mtimeMs: number;
-};
-
-/**
- * Stateless filesystem operations on absolute paths. No events, no lifecycle, no caching;
- * watching is a separate service (`IFileWatchService`).
- *
- * Intended extension set (additive, as consumers need them):
- * `list`, `write`, `glob`, `mkdir`, `copyFile`, `realPath`, `search`, `readImage`.
- * Workspace-scoping (root-relative paths + escape validation) is a wrapper layered on top,
- * not part of this interface.
- */
-export type IFsService = {
-  read(absPath: string, options?: { maxBytes?: number }): Promise<FileReadResult>;
-  stat(absPath: string): Promise<FileStat | null>;
-  remove(absPath: string, options?: { recursive?: boolean }): Promise<void>;
-  exists(absPath: string): Promise<boolean>;
-};
