+++ packages/core/src/git/git-worktree.ts
@@ -1,8 +1,9 @@
 import { createHash } from 'node:crypto';
+import fs from 'node:fs/promises';
 import path from 'node:path';
 import { err, ok, type Result, type Unsubscribe } from '@emdash/shared';
 import { ExecError, type BoundExec } from '../exec';
-import type { IFileWatchService, IFsService, WatchHandle } from '../fs';
+import type { IFileWatchService, WatchHandle } from '../fs';
 import { LiveModel } from '../lib';
 import {
   classifyCommitError,
@@ -63,13 +64,17 @@ const IMAGE_MIME_BY_EXT: Record<string, string> = {
 
 type Numstat = Map<string, { additions: number; deletions: number }>;
 
+type FileReadResult = {
+  content: string;
+  truncated: boolean;
+  totalSize: number;
+};
+
 export type GitWorktreeOptions = {
   worktree: string;
   gitDir: string;
   repository: GitRepository;
   exec: BoundExec;
-  fs: IFsService;
-  /** Injected file-watch service; disposed by the injector, not this class. */
   watcher: IFileWatchService;
   onError?: GitOnError;
 };
@@ -79,7 +84,6 @@ export class GitWorktree implements IGitWorktree {
   readonly gitDir: string;
   readonly repository: GitRepository;
   private readonly exec: BoundExec;
-  private readonly fs: IFsService;
   private readonly statusModel: LiveModel<GitStatusModel>;
   private readonly headModel: LiveModel<GitHeadModel>;
   private readonly worktreeWatch: WatchHandle;
@@ -90,7 +94,6 @@ export class GitWorktree implements IGitWorktree {
     this.gitDir = options.gitDir;
     this.repository = options.repository;
     this.exec = options.exec;
-    this.fs = options.fs;
     const onError = options.onError ?? (() => {});
 
     this.statusModel = new LiveModel<GitStatusModel>({
@@ -526,7 +529,7 @@ export class GitWorktree implements IGitWorktree {
       const deletions = unstagedNumstat.get(filePath)?.deletions ?? 0;
       if (additions === 0 && deletions === 0 && isUntracked) {
         try {
-          const result = await this.fs.read(path.join(this.worktree, filePath), {
+          const result = await readFile(path.join(this.worktree, filePath), {
             maxBytes: MAX_DIFF_CONTENT_BYTES,
           });
           if (!result.truncated) additions = (result.content.match(/\n/g) ?? []).length;
@@ -640,6 +643,37 @@ function parseNumstat(stdout: string): Numstat {
     map.set(filePath, current);
   }
   return map;
+}
+
+async function readFile(
+  filePath: string,
+  options: { maxBytes?: number } = {}
+): Promise<FileReadResult> {
+  const stat = await fs.stat(filePath);
+  const maxBytes = options.maxBytes;
+  if (maxBytes === undefined) {
+    const content = await fs.readFile(filePath, 'utf8');
+    return {
+      content,
+      truncated: false,
+      totalSize: stat.size,
+    };
+  }
+
+  const handle = await fs.open(filePath, 'r');
+  try {
+    const buffer = Buffer.alloc(maxBytes + 1);
+    const { bytesRead } = await handle.read(buffer, 0, maxBytes + 1, 0);
+    const truncated = bytesRead > maxBytes;
+    const slice = buffer.subarray(0, truncated ? maxBytes : bytesRead);
+    return {
+      content: slice.toString('utf8'),
+      truncated,
+      totalSize: stat.size,
+    };
+  } finally {
+    await handle.close();
+  }
 }
 
 function resolveDiffTarget(base: DiffTarget): { cached: boolean; ref: string } {
